moved get_adjecency_types and merge_adjacencies
