+++ src/qm9/utils.py
@@ -20,7 +20,7 @@ from combinatorial_data.combinatorial_data_utils import (
 from combinatorial_data.lifter import Lifter
 from qm9.lifts.registry import lifter_registry
 from qm9.qm9_cc import QM9_CC
-from utils import get_adjacency_types, merge_adjacencies
+# from utils import get_adjacency_types, merge_adjacencies
 
 def calc_mean_mad(loader: DataLoader) -> tuple[Tensor, Tensor]:
     """Return mean and mean average deviation of target in loader."""
